Added a model CMakeLists.txt file for OpenCV 3.1.0 projects
The CMakeLists.txt file included can be used as a template for image
projects using OpenCV 3.1.0 (from the specified location). The file
specifies the location of the OpenCV 3.1.0 configuration and
libraries (the paths are hardcoded), and ensures the correct version
(OpenCV 3.1.0) is linked to the image project.

Also added a .gitignore file to instruct git to ignore the files that
CMake creates, as these do not have to be stored in the repository.

	new file:   .gitignore
	new file:   CMakeLists.txt