Added a bash script for simple benchmarking based on execution time
The script benchmark.sh is intended to serve as a basis for our
benchmarking suite. Briefly, the script uses unix timeprogram to
time the execution of a specified process (executable file or
otherwise), which in our case will be the registration algorithm
executable. This is done in a loop, so we can get an average estimate
of the execution time, not just one-shot execution time which will
vary somewhat.

Given the implementation using unix time, and "real time" instead of
CPU time, this approach should work for all of the platforms we are
required to test.

Currently only this execution time based measure is implemented, but
other potential metrics are listed in the notes in the file.

	new file:   benchmark.sh