Work on using a Mat object to get an image to shared memory on Parallella
This is speculative work on using an OpenCV Mat object to load an image
to shared memory, to be then used on the eCores.

It is likely that this has to be made into a more "primitive" type for use
on the eCores, so probably will have to convert to a 2D array (maybe even a
1D array/vector, as the cores don't necessarily need to know the positions
along the row-column dimensions).
[This is probably the most direct approach, more likely to work than Mat.]