Added results from Epiphany bandwidth tests run on 24/2/2016
To test the memory access bandwidth of the Epiphany coprocessor
on our Parallella, I ran the bandwidth test (provided by Adapteva).
The results are saved here.

The results show the read and write speeds between combinations of
the ARM CPU, eCore (the first one in most cases), and the shared
memory. The results illustrate the slow speed of communication between
the eCores and the rest of the system.

Moreover, they indicate which type of communication is most efficient
in cases where alternative methods are available.

The results can be used in our report, to illustrate the importance of
considering the limiatations of the architecture when converting
algorithms to the Parallella, and as potential reasons for the
algorithm running slowly if this turns out to be the case.