Added a gprof sample with CMake
The "gprof_sample" folder includes a simple project [sorting, originally
by Chris Smallwood] illustrating gprof and CMake.
This is to illustrate that one can use the provided cmake, and then
run gprof as one would if it had been compiled and linked using makefile.

To run, use the shell commands:
   1. "cmake ."          [run cmake]
   2. "./sorting"        [run exe - creates gmon.out for gprof]
   3. "gprof ./sorting"  [profiles the program using gprof]

	new file:   Benchmarking/gprof_sample/CMakeLists.txt
	new file:   Benchmarking/gprof_sample/loop.sh
	new file:   Benchmarking/gprof_sample/sorting.cpp