Tidied fft_lpf code (still WIP); added a command-line argument for cpu/gpu
The fft_lpf.cpp code now runs either the cpu or gpu version based on
a command-line argument passed in.

Currently the code only does the FFT part, and the gpu version has a
bug that results in incorrect output. The cpu version works, though,
so work on implementing and testing the LPF can be made in parallel
(on the cpu first) while the gpu code for FFT is debugged.

Also included the Makefile (as this was previously ignored).

	new file:   Makefile
	deleted:    fft_lpf
	modified:   fft_lpf.cpp
	deleted:    fft_lpf.o