Adding code for FFT and LPF for testing GPU vs Parallella (WIP)
The code is still a work in progress, but it's at a stage where
parallel working on the LPF is possible while the remaining bugs
are ironed out of the cuda version of FFT.

Currently the code only does the FFT part, and the gpu version has a
bug that results in incorrect output. The cpu version works, though,
so work on implementing and testing the LPF can be made in parallel
(on the cpu first) while the gpu code for FFT is debugged.

	new file:   Makefile
	new file:   fft_lpf.cpp
	new file:   lenna.jpg