Changes to "assembly-opt" example for running it (made on 8 March)
Committing changes made on 8 March to the "assembly-opt" example
on matrix multiplication on Parallella.

The changes were made to debug the program, as it was not compiling
correctly (the example was originally made for 64-core Epiphany, and
there had been an update to the eSDK that required some further changes).

NOTE: This code is still not running correctly, as it fails to run on the
eCores for some reason (it abandons execution when the kernel code is meant
to be loaded to the cores).

	modified:   Parallella/pedagogicalExamples/assembly-opt/src/defs_multi.h
	modified:   Parallella/pedagogicalExamples/assembly-opt/src/e_dev_main.c
	modified:   Parallella/pedagogicalExamples/assembly-opt/src/host_multi.c