Testing simple image processing on Parallella; LPF-only OpenCV code (cpu)
Added a new folder for working on image processing on Parallella, at
"image_to_cores/fft-based", which is for making simple image
processing based on the FFT2D ("lena") example from the Parallella
community.

Also added a *simple* LPF-only code on OpenCV, to test if we can get
the same result as the Parallella example using filter2D with a basic
convolution-kernel based filtering. This code works (on CPU), but the
filtering does not yield the same result as the Parallella example,
and it looks like the filtering method must be very different.
Figuring out what the Parallella LPF does exactly and replicating this
is one potential avenue of work, but we shall focus now on building
something that allows us to do simple image data manipulation on the
eCores.

(Jukka and Chris)

	modified:   data_struct.h
	modified:   ../lena/device/src/fft2d_main.c
	modified:   ../lena/host/src/fft2d_host.c
	modified:   ../../../FFT and LPF/Makefile
	new file:   ../../../FFT and LPF/lpf_only.cpp

	And new files in "fft-based/"