First simple image data manipulation; debugging _USE_DRAM_ in FFT code
First simple image manipulation using our image-to-cores project based
on the FFT2D ("lena") example. The resulting image has only a few pixels
changed to different values, but it's a proof that we are working on
the right things.

	modified:   fft-based/device/Release/src/fft2d_main.o
	modified:   fft-based/device/src/fft2d_main.c
	modified:   fft-based/host/Release/matprt.m
	modified:   fft-based/lenna.lpf.jpg

On the FFT code, debugged some of the _USE_DRAM_ code, althout some
errors remain - will get back to this later.
	modified:   ../lena/build.sh
	modified:   ../lena/device/src/fft2d_main.c
	modified:   ../lena/host/src/fft2d_host.c