Version of scalar multiplication with a simple timer (WIP)
A simple timer added to fft2d_main.c around the calc() call,
to time the multiplication process (no loading overheads etc).

Also changed the multiplication to 0.5 so we don't get the
problem of going back to black (due to running out of space
for the bits).

Still need to tidy things up and improve the timer, make the
multiplier an argument rather than fixed value, and look into
different image sizes etc. (see prev commit for list of TODOs).

	modified:   device/Release/src/fft2d_main.o
	modified:   device/src/fft2d_main.c
	modified:   host/Release/matprt.m
	modified:   host/src/fft2d_host.c
	modified:   lenna.jpg
	modified:   lenna.lpf.jpg