Scalar multiply: added timers; setup passing a scalar; fixed "drop to zero"
Added timers so that the process can be timed in different ways,
including timing the memory access overhead etc.

Passing a scalar value into the scalar multiplication was
harder than one would have thought, but works.
Passing a value into the process required
1. setting up a new part to a memory structure (in fft2d.h),
2. setting up writing to and reading from this memory
3. passing a value to the host process
   - using normal argc and argv[], but also
   - passing a value to a bash script (run.sh) and onward
     to the executable (for the argv[])

Fixed the "dropping to zero" issue where multiplication could lead
to values exceeding the memory allocated (so values above 255 for
8-bit unsigned integers), thus effectively going back to zero.
This was done by just adding a check if the value would exceed 255,
and setting if to 255 if that was the case (so a white would be kept
white and not turned to black).

Tidied the code a little bit, but still a lot to do on that front.

	modified:   device/src/fft2d.h
	modified:   device/src/fft2d_main.c
	modified:   host/src/fft2d_host.c
	modified:   run.sh