Debugged the DRAM part of the Parallella image manipulation project
Debugged the DRAM sections of the Parallella image manipulation
project, so we can now use the external memory for passing the
image data to cores.
This appears to be working as it should, although the speed is
rather similar to the other approach (which should be faster).
This may be due to the small image size, though.

Whether the program is run with DRAM or local memory passing
is still controlled via pre-processor macros in the files
epiphany_main.c and host.c. This should be improved, so it
could be determined either when the project is built or run.

	modified:   device/src/epiphany_main.c
	modified:   host/src/host.c