image-to-cores: removed old timers from host.c; added 'multirun.sh'
Removed the seemingly incorrect (unstable?) timers from the host
code for the "image-to-cores"/scalar multiplication part. These
were created by Adapteva, and based on some count of cycles, but
sometimes they suggested completely unrealistic cycle counts and
runtimes.

Added a "multirun" script to run the process over a loop, for
measuring power consumption on the Parallella (as the measurement
system required a longer-running process to get the measurements to
show).

	modified:   host/src/host.c
	new file:   multirun.sh