Parallella scalar-matrix multiplication: tidying up; 64x64 works
Removed an unnecessary hard-coded line from epiphany_main.c. Not
only was it unnecessary, it was also stopping the 64x64 runs from
working. The line was added for debugging purposes, but forgotten
there.

Now the system works with 64x64 images, but one has to be careful
in the build process - if the epiphany_main.o does not get updated
when switching between image sizes, there is a problem that stops
the project working correctly. So have to ensure it gets run.

	modified:   device/src/epiphany_main.c