Added a different timer (using time.h instead of sys/time.h) to host.c
Noticed that the timers using sys/time.h are not consistent when
the program is run "normally" so that output stream is directed
to the screen, compared to when output stream is directed to a file.
The latter produces smaller time estimates.

This was a potential problem for the benchmarks, so I added a timer
implementation using time.h, as that was the method used for the
Jetson TX1 benchmarks.

The timers based on time.h seem robust to the different conditions,
and are consistent between running with the output redirected to
a file and normally directed to screen. What is more, they are
consistent with the times we used in the benchmark runs, which
were run with the output directed to a file.

SO: No changes to the benchmarkin results is required.

	modified:   host/src/host.c