Parallella: scalar_multiplication - cleaned up version, new location
Made a cleaning-up exercise of the code, as many unnecessary parts
remained. Also moved this into a better place, Parallella/scalar_mult,
as it was clearly in an unwieldy place previously.

The current state looks good, but will still have to test
that the output is correct (had trouble copying images over, for some
reason). Then can do a final cleanup, and add the instruction text
to this.

Also, this version comes with a better run.sh script, which now
only requires the image name to be inserted (while previously needed
to add "../../" in front of the name).